+++ examples/apps/web-router/src/store/index.js
@@ -27,4 +27,5 @@ export const store = createStore({
   },
   entities,
   middlewares,
+  autoCreateEntities: true,
 })