+++ benchmarks/dashboard/CHANGELOG.md
@@ -1,5 +1,16 @@
 # dashboard-benchmark
 
+## 0.2.0
+
+### Minor Changes
+
+- Remove allowTypes flag in favor of entity auto-creation
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.2.0
+
 ## 0.1.12
 
 ### Patch Changes