+++ docs/engine/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/docs
 
+## 1.10.17
+
+### Patch Changes
+
+- @inglorious/engine@20.0.0
+- @inglorious/renderer-2d@19.0.0
+- @inglorious/renderer-react-dom@19.0.0
+
 ## 1.10.16
 
 ### Patch Changes