+++ examples/apps/web-router/CHANGELOG.md
@@ -1,5 +1,16 @@
 # lit-todomvc
 
+## 0.5.0
+
+### Minor Changes
+
+- Remove allowTypes flag in favor of entity auto-creation
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.2.0
+
 ## 0.4.24
 
 ### Patch Changes