+++ examples/apps/web-select/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-select
 
+## 0.2.23
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.2.0
+
 ## 0.2.22
 
 ### Patch Changes