+++ examples/apps/web-table/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.1.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.2.0
+
 ## 0.1.33
 
 ### Patch Changes