+++ examples/apps/web-todomvc-cs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.2.0
+
 ## 0.0.26
 
 ### Patch Changes